Separate TestDatabaseCompact
