+++ jena-db/jena-tdb2/src/test/java/org/apache/jena/tdb2/sys/TS_Sys.java
@@ -25,6 +25,7 @@ import org.junit.runners.Suite;
 @RunWith(Suite.class)
 @Suite.SuiteClasses( {
     TestDatabaseOps.class
+    , TestDatabaseCompact.class
     , TestDatabaseConnection.class
     , TestSys.class
     , TestTransactionalSystemControl.class
